+++ src/mcga/brainfuck/Parser.java
@@ -1,5 +1,9 @@
 package mcga.brainfuck;
 
+import javax.imageio.ImageIO;
+import java.awt.*;
+import java.awt.image.BufferedImage;
+import java.io.IOException;
 import java.io.InputStream;
 import java.util.Scanner;
 
@@ -10,7 +14,8 @@ import java.util.Scanner;
 public abstract class Parser {
 
     protected InputStream stream;
-
+    public static final int SQUARE_SIDE = 3;
+    public static final String EMPTY_INSTRUCTION = "000000";
     /**
      * In case a file is specified in the launching command, this constructor is called.
      *
@@ -46,6 +51,51 @@ public abstract class Parser {
                 System.err.println(e.getMessage());
                 System.exit(42);
             }
+        }
+    }
+
+    public void readBitmap() {
+        try {
+            BufferedImage image = ImageIO.read(stream);
+            int height = image.getHeight();
+            int width = image.getWidth();
+            if (height % SQUARE_SIDE != 0 || width % SQUARE_SIDE != 0) {
+                throw new InvalidBitmapException();
+            }
+
+            String prevColor = "";
+            String hexColor = "";
+            for (int i = 0; i < height; i += SQUARE_SIDE) {
+                for (int j = 0; j < width; j += SQUARE_SIDE) {
+
+                    for (int iSquare = 0; iSquare < SQUARE_SIDE; iSquare++) {
+                        for (int jSquare = 0; jSquare < SQUARE_SIDE; jSquare++) {
+                            hexColor = "";
+                            Color imgColor = new Color(image.getRGB(jSquare + j, iSquare + i));
+                            hexColor += String.format("%02X", imgColor.getRed());
+                            hexColor += String.format("%02X", imgColor.getGreen());
+                            hexColor += String.format("%02X", imgColor.getBlue());
+                            if (!prevColor.equals(hexColor) && iSquare != 0 && jSquare != 0) {
+                                throw new InvalidBitmapException();
+                            }
+                            prevColor = hexColor;
+
+                        }
+                    }
+                    if (!hexColor.equals(EMPTY_INSTRUCTION)) {
+                        execute(hexColor);
+                    }
+                }
+            }
+        } catch (InvalidInstructionException e) {
+            e.printStackTrace();
+            System.exit(42);
+        } catch (InvalidBitmapException e) {
+            e.printStackTrace();
+            System.exit(14);
+        } catch (IOException e) {
+            e.printStackTrace();
+            System.exit(3);
         }
     }
