+++ src/mcga/brainfuck/Right.java
@@ -5,11 +5,11 @@ package mcga.brainfuck;
  */
 public class Right extends Movement {
     @Override
-    public void interpret() throws IndexOutOfBoundsException, InvalidValueException {
+    public void interpret() throws IndexOutOfBoundsException {
         moveR();
     }
 
-    private void moveR() throws IndexOutOfBoundsException, InvalidValueException {
+    private void moveR() throws IndexOutOfBoundsException {
         Brainfuck.memoire.changeCurrentIndex(1);
     }
